refactor: Sync controller to simple DataPod controller
